requirements.txt
streamlit~=1.39.0
transformers~=4.44.1
moviepy~=1.0.3
torchvision~=0.19.1
pillow~=10.4.0